Refactored entire application
In the process of refactoring the application, I divided the project into two main directories: client (front-end) and server (back-end). For the front-end, I reorganized the file structure to include assets, css, and scripts directories. I installed TypeScript to enhance JavaScript code with types, and TailwindCSS for utility-first styling. This allows for a more modular and maintainable codebase.

On the server side, I established a TypeScript environment and set up an Express server with Node.js. I created a server.ts as the entry point for the server application. For cross-origin resource sharing, I configured CORS to allow the front-end to communicate with the back-end securely. I also installed Prisma as an ORM for PostgreSQL, which I then connected to a hosted database on Railway, ensuring persistent data storage and management. The server directory now includes specific folders for controllers, routes, and utilities.