Completed database population with 250 countries.
I developed and executed utility functions to fetch and process country and GDP data, then populated the database with this information. Initially, I encountered issues with missing countries, which I resolved by creating an additional function to identify, fetch, and insert the missing entries. This process, primarily a one-time requirement due to the static nature of the data, successfully added all 250 countries to the Prisma database.

The next step will be to create front and back-end logic for fetching the Country data from the database so that it can be displayed.