Fixed bug to set default alphabetical sorting.
A bug was identified in the country sorting functionality of the application. The expected behavior was for the country list to be sorted alphabetically upon initial page load. However, this was not occurring due to the default filter state not being explicitly set for alphabetical sorting. To rectify this, the sortCountries function was updated to include a specific case for 'alphabetical' sorting. Additionally, the currentFilters.misc was set to 'alphabetical' within the main.ts file to establish this as the default sorting method. Now, when the application loads, the country list is automatically sorted alphabetically without requiring user interaction to select this option.